requirements.txt
django~=4.2.6
django_extensions~=3.2.3
django-debug-toolbar~=4.2.0
mozilla-django-oidc~=3.0.0
djangorestframework~=3.14.0
PyYAML~=6.0.1
uritemplate~=4.1.1
Markdown~=3.5
Pygments~=2.16.1
django-filter~=23.3
django-guardian~=2.4.0
psycopg-binary~=3.1.12
gunicorn~=21.2.0
whitenoise~=6.6.0
